+++ cmd/server-manager/typescript/src/RaceControl.ts
@@ -689,6 +689,7 @@ class LiveTimings implements WebsocketHandler {
             // be removed from the disconnected drivers table and placed into the connected drivers table.
             const dummyDriver = JSON.parse(JSON.stringify(driver));
             dummyDriver.CarInfo.CarModel = carName;
+            dummyDriver.CarInfo.CarName = driver.Cars[carName].CarName;
 
             this.addDriverToTable(dummyDriver, this.$disconnectedDriversTable);
         }
@@ -799,7 +800,7 @@ class LiveTimings implements WebsocketHandler {
         }
 
         // car model
-        $tr.find(".driver-car").text(driver.CarInfo.CarName ? driver.CarInfo.CarName : prettifyName(driver.CarInfo.CarModel, true));
+        $tr.find(".driver-car").text(carInfo.CarName ? carInfo.CarName : prettifyName(driver.CarInfo.CarModel, true));
 
         if (addingDriverToConnectedTable) {
             let currentLapTimeText = "";